INTERLOK-3447 Abstraction + HTTP status propogation
Refactor to have a GenericAccessTokenImpl for use with INTERLOK-3440
Propagate the HTTP status via a Consumer that gets passed through.
Refactor the tests to reflect the changes.
